+++ src/components/User/Users/components/UserDetails/UserDetails.Request.vue
@@ -2,106 +2,11 @@
   <div class="user-request">
     <template v-if="requestToReview.isPending">
       <div class="user-details-request__manage-tasks-section">
-        <h3 class="user-details-request__manage-tasks-title">
-          Manage tasks (advanced)
-        </h3>
-
-        <h4 class="user-details-request__task-section-heading">
-          Current state
-        </h4>
-
-        <div class="user-details-request__task-section-content">
-          <p>Pending tasks: {{ requestToReview.pendingTasks }}</p>
-          <p>All tasks: {{ requestToReview.allTasks }}</p>
-        </div>
-
-        <h4 class="user-details-request__task-section-heading">
-          Tasks to add
-        </h4>
-
-        <div class="user-details-request__task-section-content">
-          <!--eslint-disable max-len-->
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToAdd"
-            :label="`Submit auto verification request (${CHANGE_ROLE_TASKS.submitAutoVerification})`"
-            :cb-value="CHANGE_ROLE_TASKS.submitAutoVerification"
-            :disabled="isPending"
-          />
-          <!--eslint-enable max-len-->
-
-          <!--eslint-disable max-len-->
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToAdd"
-            :label="`Complete auto verification request (${CHANGE_ROLE_TASKS.completeAutoVerification})`"
-            :cb-value="CHANGE_ROLE_TASKS.completeAutoVerification"
-            :disabled="isPending"
-          />
-          <!--eslint-enable max-len-->
-
-          <!--eslint-disable max-len-->
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToAdd"
-            :label="`Manual review required (${CHANGE_ROLE_TASKS.manualReviewRequired})`"
-            :cb-value="CHANGE_ROLE_TASKS.manualReviewRequired"
-            :disabled="isPending"
-          />
-          <!--eslint-enable max-len-->
-
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToAdd"
-            :label="`Default (${CHANGE_ROLE_TASKS.default})`"
-            :cb-value="CHANGE_ROLE_TASKS.default"
-            :disabled="isPending"
-          />
-        </div>
-
-        <h4 class="user-details-request__task-section-heading">
-          Tasks to remove
-        </h4>
-
-        <div class="user-details-request__task-section-content">
-          <!--eslint-disable max-len-->
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToRemove"
-            :label="`Submit auto verification request (${CHANGE_ROLE_TASKS.submitAutoVerification})`"
-            :cb-value="CHANGE_ROLE_TASKS.submitAutoVerification"
-            :disabled="isPending"
-          />
-          <!--eslint-enable max-len-->
-
-          <!--eslint-disable max-len-->
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToRemove"
-            :label="`Complete auto verification request (${CHANGE_ROLE_TASKS.completeAutoVerification})`"
-            :cb-value="CHANGE_ROLE_TASKS.completeAutoVerification"
-            :disabled="isPending"
-          />
-          <!--eslint-enable max-len-->
-
-          <!--eslint-disable max-len-->
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToRemove"
-            :label="`Manual review required (${CHANGE_ROLE_TASKS.manualReviewRequired})`"
-            :cb-value="CHANGE_ROLE_TASKS.manualReviewRequired"
-            :disabled="isPending"
-          />
-          <!--eslint-enable max-len-->
-
-          <tick-field
-            class="app__form-field"
-            v-model="tasksToRemove"
-            :label="`Default (${CHANGE_ROLE_TASKS.default})`"
-            :cb-value="CHANGE_ROLE_TASKS.default"
-            :disabled="isPending"
-          />
-        </div>
+        <tasks-manager
+          v-model="tasks"
+          :request="requestToReview"
+          :is-pending="isPending"
+        />
       </div>
 
       <div class="app__form-actions user-details-request__form-actions">
@@ -173,9 +78,10 @@
 
 <script>
 import api from '@/api'
-import config from '@/config'
 
-import { TextField, TickField } from '@comcom/fields'
+import TasksManager from './UserDetails.TasksManager'
+
+import { TextField } from '@comcom/fields'
 import Modal from '@comcom/modals/Modal'
 
 import 'mdi-vue/ChevronDownIcon'
@@ -195,7 +101,7 @@ export default {
   components: {
     Modal,
     TextField,
-    TickField,
+    TasksManager,
   },
 
   props: {
@@ -215,22 +121,14 @@ export default {
 
   data () {
     return {
-      tasksToAdd: 0,
-      tasksToRemove: 0,
-
       rejectForm: {
         reason: '',
         isShown: false,
       },
       isShownAdvanced: false,
       isPending: false,
-
-      CHANGE_ROLE_TASKS: config.CHANGE_ROLE_TASKS,
+      tasks: {},
     }
-  },
-
-  async created () {
-    this.tasksToRemove = this.requestToReview.pendingTasks
   },
 
   methods: {
@@ -241,8 +139,8 @@ export default {
       this.isPending = true
       try {
         const reviewDetails = this.requestToReview.record.reviewDetails || {}
-        reviewDetails.tasksToAdd = this.tasksToAdd
-        reviewDetails.tasksToRemove = this.tasksToRemove
+        reviewDetails.tasksToAdd = this.tasks.toAdd
+        reviewDetails.tasksToRemove = this.tasks.toRemove
 
         await api.requests.approve({
           ...this.requestToReview.record,
@@ -301,17 +199,5 @@ export default {
 <style scoped>
 .user-request__btn {
   width: 15rem;
-}
-
-.user-details-request__manage-tasks-title {
-  margin-bottom: 1.5rem;
-}
-
-.user-details-request__task-section-heading {
-  margin-bottom: 1rem;
-}
-
-.user-details-request__task-section-content {
-  margin-bottom: 2rem;
 }
 </style>