+++ CHANGELOG.md
@@ -16,6 +16,10 @@ for further information about branching and tagging conventions.
 - Add link to pre-issuance guide
 - Show all the limits in one screen, "Not set" and "Unlimited" labels,
   ability to remove limits and move `Limits.Manager` to v3 endpoint
+- KYC verification queue page for quick requests review
+- "US Verified" & "US Accredited" roles to the roleIdToString filter
+- "hash", "type", & "requestor" fields to `ChangeRoleRequest` wrapper
+- PDF documents preview to the `UserDocGetter` component
 
 #### Fixed
 - Fixed a bug when we confirm a request for changes in limits,
@@ -24,6 +28,9 @@ for further information about branching and tagging conventions.
 - Displaying of confirmation modal on the issuance form
 - Uploading documents on create asset form
 - Asset select in withdrawal list
+
+#### Updated
+- Extracted change role request tasks manager to a separate component
 
 ## [1.6.0] - 2019-05-09
 ### "Under the hood" changes